refactor(BlackJackGame): 도메인으로 파일위치 이동
